backend/requirements.txt
flask
flask-cors
gunicorn
google-genai
supabase
python-dotenv